feat(JavaRush): converter

+++ src/com/javarush/task/pro/task09/task0902/Solution.java
@@ -5,11 +5,11 @@ package com.javarush.task.pro.task09.task0902;
 */
 
 public class Solution {
-    public int a = 0.;
-    public int b = .0;
-    public int c = 100D;
-    public int d = 100.0;
-    public int e = 1.11E5;
-    public int f = 200F;
-    public int g = 0.F;
+    public double a = 0.;
+    public double b = .0;
+    public double c = 100D;
+    public double d = 100.0;
+    public double e = 1.11E5;
+    public float f = 200F;
+    public float g = 0.F;
 }
